use mysql instead of postgres

+++ product/src/main/resources/db/migration/V001__00_create_table_product.sql
@@ -1,7 +1,7 @@
 create table product (
     id serial primary key,
     name varchar(255),
-    price float8
+    price float
 );
 
 insert into product(name, price) values ('product1', 21.2);
